Partition at rules before building context
